+++ src/types/photo.ts
